fix exception handling
